+++ packages/h3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/h3
 
+## 0.1.89
+
+### Patch Changes
+
+- @vercel/node@5.8.4
+
 ## 0.1.88
 
 ### Patch Changes