+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/nestjs
 
+## 0.2.84
+
+### Patch Changes
+
+- @vercel/node@5.8.4
+
 ## 0.2.83
 
 ### Patch Changes