+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/static-build
 
+## 2.9.30
+
+### Patch Changes
+
+- @vercel/gatsby-plugin-vercel-builder@2.2.7
+
+## 2.9.29
+
+### Patch Changes
+
+- 23eee91: Revert "[static-build] Support Vite Environments API for SSR projects (#16354)"
+
 ## 2.9.28
 
 ### Patch Changes